fix(cache): check file exists

+++ utils.py
@@ -88,7 +88,8 @@ def update_settings(data):
         config_dict["server"]["host"] = data["host"]
         config_dict["server"]["port"] = int(data["port"])
 
-    os.remove('server_config.json')
+    if os.path.exists('server_config.json'):
+        os.remove('server_config.json')
 
     with open('config.json', 'w') as config:
         json.dump(config_dict, config)
